This commit includes two files main files:
1) networks.py
2) training.py

networks.py has two networks, a DCGAN Generator, and
a discriminator implementing mini-batch discrimination.
Several other networks are also included but they are
intentionally commented out.

training.py trains the GAN to recognize MNIST digits.
A one line script at the bottom of the file shows an
example of how to use it.